Separate websocket event dispatch from subscription lifecycle
The websocket manager mixed connection lifecycle, event dispatch, and cache invalidation logic in one large file. Split the implementation into subscription, dispatch, and shared type modules while preserving the existing WebSocketManager public surface for callers and tests.

Constraint: Keep the adapter import contract stable for existing callers and focused tests.
Rejected: Rename WebSocketManager imports across callers | unnecessary scope expansion for this extraction.
Confidence: high
Scope-risk: narrow
Tested: deno check src/platform/adapters/fs/veryfront/websocket-manager.test.ts src/platform/adapters/fs/veryfront/websocket-manager.ts src/platform/adapters/fs/veryfront/websocket/subscription.ts src/platform/adapters/fs/veryfront/websocket/dispatch.ts src/platform/adapters/fs/veryfront/websocket/types.ts
Tested: deno lint src/platform/adapters/fs/veryfront/websocket-manager.test.ts src/platform/adapters/fs/veryfront/websocket-manager.ts src/platform/adapters/fs/veryfront/websocket/types.ts src/platform/adapters/fs/veryfront/websocket/subscription.ts src/platform/adapters/fs/veryfront/websocket/dispatch.ts
Tested: deno test --no-check --allow-all src/platform/adapters/fs/veryfront/websocket-manager.test.ts
Tested: Pre-commit verify:quick hook passed
Tested: LSP diagnostics on modified websocket files reported 0 errors
